requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure Functions
azure-functions==1.17.0

# Application Insights
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
applicationinsights==0.11.10

# Azure OpenAI Integration
openai>=1.0.0
azure-identity>=1.14.1
azure-keyvault-secrets>=4.7.0

# Database - Required for Find Course API
asyncpg==0.29.0
pgvector==0.3.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0  # For parallel test execution
pytest-timeout==2.2.0  # For test timeout management
aiohttp==3.12.13  # For performance testing

# Development
python-dotenv==1.0.0
nest-asyncio==1.6.0  # For handling nested event loops in tests

# Logging and Monitoring
python-json-logger==2.0.7
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
pytz==2024.1

# Configuration
pyyaml>=6.0.1

# Language Detection
langdetect==1.0.9

# Text Processing and ML
beautifulsoup4==4.12.2
numpy==1.24.3
scikit-learn==1.3.2

# Rate Limiting (Optional)
# slowapi==0.1.9

# Development Tools
ruff==0.5.7
